feat: fix bugs

+++ src/modules/devices/service/devices.service.ts
@@ -10,12 +10,14 @@ import jwt from 'jsonwebtoken';
 import { JWT } from '../../auth/constants';
 import { JwtRepository } from '../../auth/repository/jwt.repository';
 import { DeleteResult } from 'mongodb';
+import { JwtService } from '@nestjs/jwt';
 
 @Injectable()
 export class DevicesService {
   constructor(
     public devicesRepository: DevicesRepository,
     public jwtRepository: JwtRepository,
+    public jwtService: JwtService,
   ) {}
 
   async createUserSession(
@@ -66,6 +68,14 @@ export class DevicesService {
     const getUserDataByToken = await this.tokenVerify(refreshToken);
     if (!getUserDataByToken) throw new UnauthorizedException([]);
     const userId = getUserDataByToken.userId;
+    const getLastActiveDate = await this.getLastActiveDate(refreshToken);
+    const getUserSession =
+      await this.devicesRepository.findOneByDeviceIdUserIdAndLastActiveDate(
+        userId,
+        getUserDataByToken.deviceId,
+        getLastActiveDate,
+      );
+    if (!getUserSession) throw new UnauthorizedException([]);
     return this.devicesRepository.findAllUserDevicesByUserId(userId);
   }
 
@@ -77,7 +87,10 @@ export class DevicesService {
       return null;
     }
   }
-
+  async getLastActiveDate(refreshToken: string): Promise<string> {
+    const payload: any = await this.jwtService.decode(refreshToken);
+    return new Date(payload.iat * 1000).toISOString();
+  } // 1
   async deleteAllDevices(refreshToken: string): Promise<DeleteResult> {
     // const findUserInBlackList =
     //   await this.jwtRepository.findRefreshTokenInBlackList(refreshToken);
@@ -86,7 +99,16 @@ export class DevicesService {
     if (!getUserDataByToken) throw new UnauthorizedException([]);
     const userId = getUserDataByToken.userId;
     const deviceId = getUserDataByToken.deviceId;
-    console.log('deviceId', typeof deviceId);
+    //
+    const getLastActiveDate = await this.getLastActiveDate(refreshToken);
+    const getUserSession =
+      await this.devicesRepository.findOneByDeviceIdUserIdAndLastActiveDate(
+        userId,
+        getUserDataByToken.deviceId,
+        getLastActiveDate,
+      );
+    //
+    if (!getUserSession) throw new UnauthorizedException([]);
     return this.devicesRepository.deleteAllDevicesById(userId, deviceId);
   }
 
@@ -102,6 +124,16 @@ export class DevicesService {
     if (!findDeviceByDeviceId) throw new NotFoundException(['User not found']);
     if (getUserDataByToken.userId !== findDeviceByDeviceId.userId)
       throw new ForbiddenException(['Its not your device']);
+    //
+    const getLastActiveDate = await this.getLastActiveDate(refreshToken);
+    const getUserSession =
+      await this.devicesRepository.findOneByDeviceIdUserIdAndLastActiveDate(
+        findDeviceByDeviceId.userId,
+        getUserDataByToken.deviceId,
+        getLastActiveDate,
+      );
+    //
+    if (!getUserSession) throw new UnauthorizedException([]);
     return this.devicesRepository.deleteUserSessionByUserAndDeviceId(
       getUserDataByToken.userId,
       deviceId,
